+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.5.2 ###
++ Added a new config option "CantOpenStorageCloset" in section "Shrunken"
+*This will prevent shrunken players from opening and closing the storage closet door.. leaving those stuck, that weren't a great asset..*
+
 ### 0.5.1 ###
 + fixed shrinking/enlarging not working when hitting exactly 0.2 or MaximumPlayerSize
 + added new config options to description